Reafactor: type interface 이동
